make `--write` only write on change
Currently, even if there are no changes `--write` will still write to
file, changing its last modified timestamp. This causes a problem for
things that monitors file modification by that timestamp.

For example treefmt `--fail-on-change` will fail, since it expects
formatters to not write to file if nothing changes:
https://github.com/numtide/treefmt/blob/main/docs/site/reference/formatter-spec.md#2-write-to-changed-files

This commit changes the behavior to only write if there are changes.